backend/requirements.txt
Flask==2.3.2
Flask-SQLAlchemy==3.0.5
Flask-Cors==4.0.0
PyJWT==2.7.0
python-dotenv==1.0.0
Werkzeug==2.3.6